Make UuidProvider usable with both ILoggingEvent and IAccessEvent
Refactor UuidProvider and associate utility classes to make them usable for both ILoggingEvent and IAccessEvent:
- create a new UuidJsonProvider by making a copy of the existing UuidProvider (restricted to ILoggingEvent)
- deprecate the existing UuidProvider in favour of the common UuidJsonProvider
- move the LoggingEventJsonProviders#addUuid() method into the common JsonProviders class

Closes #662
