+++ src/main/java/net/logstash/logback/composite/loggingevent/LoggingEventJsonProviders.java
@@ -96,9 +96,6 @@ public class LoggingEventJsonProviders extends JsonProviders<ILoggingEvent> {
     public void addNestedField(LoggingEventNestedJsonProvider provider) {
         addProvider(provider);
     }
-    public void addUuid(UuidProvider provider) {
-        addProvider(provider);
-    }
     public void addThrowableClassName(ThrowableClassNameJsonProvider provider) {
         addProvider(provider);
     }
