+++ src/main/java/net/logstash/logback/composite/loggingevent/UuidProvider.java
@@ -15,30 +15,25 @@
  */
 package net.logstash.logback.composite.loggingevent;
 
-import java.io.IOException;
-
-import net.logstash.logback.composite.AbstractFieldJsonProvider;
-import net.logstash.logback.composite.JsonWritingUtils;
+import net.logstash.logback.composite.UuidJsonProvider;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.uuid.EthernetAddress;
-import com.fasterxml.uuid.Generators;
-import com.fasterxml.uuid.NoArgGenerator;
-import com.fasterxml.uuid.impl.TimeBasedGenerator;
 
 /**
  * Outputs random UUID as field value.
  * Handy when you want to provide unique identifier for log lines.
+ * 
+ * @Deprecated use {@link UuidJsonProvider} instead.
  */
-public class UuidProvider extends AbstractFieldJsonProvider<ILoggingEvent> {
+@Deprecated
+public class UuidProvider extends UuidJsonProvider<ILoggingEvent> {
 
-    public static final String FIELD_UUID = "uuid";
+    public static final String FIELD_UUID = UuidJsonProvider.FIELD_UUID;
 
     /**
      * Type 4 UUID.
      */
-    public static final String STRATEGY_RANDOM = "random";
+    public static final String STRATEGY_RANDOM = UuidJsonProvider.STRATEGY_RANDOM;
 
     /**
      * Type 1 time based UUID.
@@ -47,70 +42,5 @@ public class UuidProvider extends AbstractFieldJsonProvider<ILoggingEvent> {
      * {@link #ethernet} can be set to either 'interface' (to automatically pick a MAC address from a network interface)
      * or a MAC address string.
      */
-    public static final String STRATEGY_TIME = "time";
-
-    private NoArgGenerator uuids = Generators.randomBasedGenerator();
-
-    /**
-     * One of {@value #STRATEGY_RANDOM} or {@value #STRATEGY_TIME}.
-     */
-    private String strategy = STRATEGY_RANDOM;
-
-    /**
-     * For {@link UuidStrategy#time} strategy only,
-     * 'interface' or ethernet MAC address.
-     */
-    private String ethernet;
-
-    public UuidProvider() {
-        setFieldName(FIELD_UUID);
-    }
-
-    @Override
-    public void writeTo(JsonGenerator generator, ILoggingEvent iLoggingEvent) throws IOException {
-        JsonWritingUtils.writeStringField(generator, getFieldName(), uuids.generate().toString());
-    }
-
-    public String getStrategy() {
-        return strategy;
-    }
-
-    public void setStrategy(String strategy) {
-        this.strategy = strategy;
-
-        uuids = newUuidStrategy(strategy, ethernet);
-    }
-
-    public String getEthernet() {
-        return ethernet;
-    }
-
-    public void setEthernet(String ethernet) {
-        this.ethernet = ethernet;
-
-        uuids = newUuidStrategy(this.strategy, this.ethernet);
-    }
-
-    private NoArgGenerator newUuidStrategy(String strategy, String ethernet) {
-
-        if (STRATEGY_TIME.equalsIgnoreCase(strategy)) {
-            return newTimeBasedGenerator(ethernet);
-        }
-        if (STRATEGY_RANDOM.equalsIgnoreCase(strategy)) {
-            return Generators.randomBasedGenerator();
-        }
-        throw new IllegalArgumentException("Unknown strategy: " + strategy);
-    }
-
-    private TimeBasedGenerator newTimeBasedGenerator(String ethernet) {
-        if (ethernet == null) {
-            return Generators.timeBasedGenerator();
-        }
-
-        if ("interface".equalsIgnoreCase(ethernet)) {
-            return Generators.timeBasedGenerator(EthernetAddress.fromInterface());
-        }
-
-        return Generators.timeBasedGenerator(EthernetAddress.valueOf(ethernet));
-    }
+    public static final String STRATEGY_TIME = UuidJsonProvider.STRATEGY_TIME;
 }
